+++ src/Bk.APG.UI/src/app/app.component.ts
@@ -58,19 +58,18 @@ export class AppComponent implements OnInit {
                         label: 'navigation.exports.dataAnalysis',
                         isExternal: false,
                     },
-                    // TODO, to be implemented in 2026
                     {
                         id: 'requestsAndReports',
                         url: 'requestsAndReports',
                         label: 'navigation.exports.requestsAndReports',
                         isExternal: false,
                     },
-                    // {
-                    //     id: 'formLetters',
-                    //     url: 'formLetters',
-                    //     label: 'navigation.exports.formLetters',
-                    //     isExternal: false,
-                    // },
+                    {
+                        id: 'formLetters',
+                        url: 'formLetters',
+                        label: 'navigation.exports.formLetters',
+                        isExternal: false,
+                    },
                 ],
             },
         ];